backend/requirements.txt
aiosmtplib==4.0.2
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.31.0
autoflake==2.3.1
bcrypt==5.0.0
black==25.12.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
croniter==6.0.0
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.124.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
isort==7.0.0
librt==0.7.3
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.3
minio==7.2.20
mypy==1.19.0
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.5.1
pluggy==1.6.0
psycopg[binary]==3.2.13
pyasn1==0.6.1
pycparser==2.23
pycryptodome==3.23.0
PyJWT==2.10.1
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
reportlab==4.4.10
requests==2.32.5
rq==2.6.1
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.44
starlette==0.50.0
stripe==14.0.1
structlog==25.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
win32_setctime==1.2.0; platform_system == "Windows"
